Selenium-python/requirements.txt
selenium==3.141
pytest==6.2
python-dotenv